WIP - investigate replacing slugified names as ids.
People's names change, so slugifying them doesn't really cut it
for the id field. However, there is nothing else which looks
particularly useful. The 'id' in the urls is certainly not a person
id - each person can have more than one. This is true whether
we work with the numerical part alone, or the slug part, or both
parts together.