Try using the slug part of the id from the querystring.
People's names change, so slugifying them doesn't really cut it
for the id field. However, there is nothing else which looks
particularly useful.

The 'id' in the urls is certainly not a person id - each person
can have more than one. This is true whether we work with the
numerical part alone, or the slug part, or both parts together.

The best compromise for now seems to be to take the slug part of
this querystring parameter, remove the term id from the end, and
use that. Some matching up of people with more than one ID will
be necessary for downstream users - sorry!